+++ appcli/orchestrators.py
@@ -44,14 +44,16 @@ class Orchestrator:
     """
 
     def start(
-        self, cli_context: CliContext, service_name: tuple[str, ...] = None
+        self,
+        cli_context: CliContext,
+        service_names: tuple[str, ...] = None,
     ) -> CompletedProcess:
         """
         Starts Docker containers (services). Optionally accepts a tuple of service names to start.
 
         Args:
-            cli_context (CliContext): The current CLI context.
-            service_names (tuple[str,...], optional): Names of the services to start. If not provided, starts all services.
+            cli_context: The current CLI context.
+            service_names: Names of the services to start. If not provided, starts all services.
 
         Returns:
             CompletedProcess: Result of the orchestrator command.
@@ -59,14 +61,16 @@ class Orchestrator:
         raise NotImplementedError
 
     def shutdown(
-        self, cli_context: CliContext, service_name: tuple[str, ...] = None
+        self,
+        cli_context: CliContext,
+        service_names: tuple[str, ...] = None,
     ) -> CompletedProcess:
         """
         Stops Docker containers (services). Optionally accepts a tuple of service names to shutdown.
 
         Args:
-            cli_context (CliContext): The current CLI context.
-            service_names (tuple[str,...], optional): Names of the services to shutdown. If not provided, shuts down all services.
+            cli_context: The current CLI context.
+            service_names: Names of the services to shutdown. If not provided, shuts down all services.
 
         Returns:
             CompletedProcess: Result of the orchestrator command.
@@ -74,14 +78,18 @@ class Orchestrator:
         raise NotImplementedError
 
     def status(
-        self, cli_context: CliContext, service_name: tuple[str, ...] = None
+        self,
+        cli_context: CliContext,
+        service_names: tuple[str, ...] = None,
+        extra_args: tuple[str, ...] = None,
     ) -> CompletedProcess:
         """
         Gets the status of Docker containers (services). Optionally accepts a tuple of service names to get the status of.
 
         Args:
-            cli_context (CliContext): The current CLI context.
-            service_names (tuple[str,...], optional): Names of the services to get the status of. If not provided, gets the status of all services.
+            cli_context: The current CLI context.
+            service_names: Names of the services to get the status of. If not provided, gets the status of all services.
+            extra_args: Additional arguments to be passed to the command.
 
         Returns:
             CompletedProcess: Result of the orchestrator command.
@@ -100,10 +108,10 @@ class Orchestrator:
         upon completing a short-lived task.
 
         Args:
-            cli_context (CliContext): The current CLI context.
-            service_name (str): Name of the container to run.
-            extra_args (Iterable[str]): Extra arguments for running the container.
-            detached (bool): Optional - defaults to False. Whether the task should run in `--detached` mode or not.
+            cli_context: The current CLI context.
+            service_name: Name of the container to run.
+            extra_args: Extra arguments for running the container.
+            detached: Optional - defaults to False. Whether the task should run in `--detached` mode or not.
 
         Returns:
             CompletedProcess: Result of the orchestrator command.
@@ -122,11 +130,11 @@ class Orchestrator:
         Executes a command in a running container.
 
         Args:
-            cli_context (CliContext): The current CLI context.
-            service_name (str): Name of the container to be acted upon.
-            command (str): The command to be executed, along with any arguments.
-            stdin_input (str): Optional - defaults to None. String passed through to the stdin of the exec command.
-            capture_output (bool): Optional - defaults to False. True to capture stdout/stderr for the run command.
+            cli_context: The current CLI context.
+            service_name: Name of the container to be acted upon.
+            command: The command to be executed, along with any arguments.
+            stdin_input: Optional - defaults to None. String passed through to the stdin of the exec command.
+            capture_output: Optional - defaults to False. True to capture stdout/stderr for the run command.
 
         Returns:
             CompletedProcess: Result of the orchestrator command.
@@ -138,7 +146,7 @@ class Orchestrator:
         Returns a click command which streams logs for Docker containers.
 
         Args:
-            cli_context (CliContext): The current CLI context.
+            cli_context: The current CLI context.
 
         Returns:
             click.Command: Command for streaming logs.
@@ -170,8 +178,8 @@ class Orchestrator:
         Checks whether a list of named services exist. Returns True if all services exist, otherwise returns False.
 
         Args:
-            cli_context (CliContext): The current CLI context.
-            service_names (tuple[str,...]): Names of the services.
+            cli_context: The current CLI context.
+            service_names: Names of the services.
 
         Returns:
             bool: if the services exists.
@@ -209,14 +217,14 @@ class DockerComposeOrchestrator(Orchestrator):
         Creates a new instance of an orchestrator for docker-compose-based applications.
 
         Args:
-            docker_compose_file (Path): Path to a docker compose file containing long-running services. Path is relative
+            docker_compose_file: Path to a docker compose file containing long-running services. Path is relative
                 to the generated configuration directory.
-            docker_compose_override_directory (Path, optional): Path to a directory containing any additional
+            docker_compose_override_directory: Optional - Path to a directory containing any additional
                 docker-compose override files. Overrides are applied in alphanumeric order of filename. Path is relative
                 to the generated configuration directory.
-            docker_compose_task_file (Path): Path to a docker compose file containing services to be run as short-lived
+            docker_compose_task_file: Path to a docker compose file containing services to be run as short-lived
                 tasks. Path is relative to the generated configuration directory.
-            docker_compose_task_override_directory (Path): Path to a directory containing any additional
+            docker_compose_task_override_directory: Path to a directory containing any additional
                 docker-compose override files for services used as tasks. Path is relative to the generated
                 configuration directory.
         """
@@ -230,7 +238,7 @@ class DockerComposeOrchestrator(Orchestrator):
     def start(
         self, cli_context: CliContext, service_names: tuple[str, ...] = None
     ) -> CompletedProcess:
-        command = ("up", "-d")
+        command = ("up", "-d", "--wait")
         if service_names is not None and len(service_names) > 0:
             command += service_names
         return self.__compose_service(cli_context, command)
@@ -247,11 +255,20 @@ class DockerComposeOrchestrator(Orchestrator):
         return self.__compose_service(cli_context, ("down",))
 
     def status(
-        self, cli_context: CliContext, service_names: tuple[str, ...] = None
+        self,
+        cli_context: CliContext,
+        service_names: tuple[str, ...] = None,
+        extra_args: tuple[str, ...] = None,
     ) -> CompletedProcess:
         command = ("ps", "-a")
         if service_names is not None and len(service_names) > 0:
             command += service_names
+        if extra_args:
+            command += extra_args
+            logger.info(
+                f"Extra arguments provided to STATUS command: [{' '.join(extra_args)}]"
+            )
+            return self.__compose_service(cli_context, command, capture_output=True)
         return self.__compose_service(cli_context, command)
 
     def task(
@@ -408,14 +425,14 @@ class DockerSwarmOrchestrator(Orchestrator):
         Creates a new instance of an orchestrator for docker swarm applications.
 
         Args:
-            docker_compose_file (Path): Path to a docker compose file containing long-running services. Path is relative
+            docker_compose_file: Path to a docker compose file containing long-running services. Path is relative
                 to the generated configuration directory.
-            docker_compose_override_directory (Path, optional): Path to a directory containing any additional
+            docker_compose_override_directory: Optional - Path to a directory containing any additional
                 docker-compose override files. Overrides are applied in alphanumeric order of filename. Path is relative
                 to the generated configuration directory.
-            docker_compose_task_file (Path): Path to a docker compose file containing services to be run as short-lived
+            docker_compose_task_file: Path to a docker compose file containing services to be run as short-lived
                 tasks. Path is relative to the generated configuration directory.
-            docker_compose_task_override_directory (Path): Path to a directory containing any additional
+            docker_compose_task_override_directory: Path to a directory containing any additional
                 docker-compose override files for services used as tasks. Path is relative to the generated
                 configuration directory.
         """
@@ -467,8 +484,12 @@ class DockerSwarmOrchestrator(Orchestrator):
         return self.__docker_stack(cli_context, ("rm",))
 
     def status(
-        self, cli_context: CliContext, service_names: tuple[str, ...] = None
+        self,
+        cli_context: CliContext,
+        service_names: tuple[str, ...] = None,
+        extra_args: tuple[str, ...] = None,
     ) -> CompletedProcess:
+        # NOTE: extra_args not yet implemented for this orchestrator.
         if service_names is not None and len(service_names) > 0:
             logger.error(
                 "Docker Swarm orchestrator cannot check the status of individual services. Attempted to get the status of [%s].",
@@ -619,8 +640,8 @@ class HelmOrchestrator(Orchestrator):
         NOTE: All `Path` objects are relative to the configuration directory.
 
         Args:
-            chart_location (Path): Path to the helm chart file/directory to deploy.
-            helm_set_values_dir (Path): The directory containing all main `values.yaml` files.
+            chart_location: Path to the helm chart file/directory to deploy.
+            helm_set_values_dir: The directory containing all main `values.yaml` files.
                 Defaults to: `${GENERATED_CONFIGURATION_DIR}/cli/helm/set-values`
 
                 All files in this directory are applied with:
@@ -628,7 +649,7 @@ class HelmOrchestrator(Orchestrator):
 
                 See below for more details.
 
-            helm_set_files_dir (Path): The directory containing all key-specific files.
+            helm_set_files_dir: The directory containing all key-specific files.
                 Defaults to: `${GENERATED_CONFIGURATION_DIR}/cli/helm/set-files`
 
                 Take the following directory:
@@ -660,13 +681,14 @@ class HelmOrchestrator(Orchestrator):
         self.helm_set_files_dir = helm_set_files_dir
 
     def start(
-        self, cli_context: CliContext, service_name: tuple[str, ...] = None
+        self, cli_context: CliContext, service_names: tuple[str, ...] = None
     ) -> CompletedProcess:
         """
         Installs (or upgrades) a helm chart inside the current kubernetes cluster.
 
         Args:
-            cli_context (CliContext): The current CLI context.
+            cli_context: The current CLI context.
+            service_names: Optional - Names of the services to get the status of. If not provided, gets the status of all services.
 
         Returns:
             CompletedProcess: Result of the orchestrator command.
@@ -713,13 +735,14 @@ class HelmOrchestrator(Orchestrator):
         return self.__run_command(command)
 
     def shutdown(
-        self, cli_context: CliContext, service_name: tuple[str, ...] = None
+        self, cli_context: CliContext, service_names: tuple[str, ...] = None
     ) -> CompletedProcess:
         """
         Uninstalls a helm chart from current kubernetes cluster.
 
         Args:
-            cli_context (CliContext): The current CLI context.
+            cli_context: The current CLI context.
+            service_names: Optional - Names of the services to get the status of. If not provided, gets the status of all services.
 
         Returns:
             CompletedProcess: Result of the orchestrator command.
@@ -737,17 +760,23 @@ class HelmOrchestrator(Orchestrator):
         return self.__run_command(command)
 
     def status(
-        self, cli_context: CliContext, service_name: tuple[str, ...] = None
+        self,
+        cli_context: CliContext,
+        service_names: tuple[str, ...] = None,
+        extra_args: tuple[str, ...] = None,
     ) -> CompletedProcess:
         """
         Get the status of the chart (through `helm status`).
 
         Args:
-            cli_context (CliContext): The current CLI context.
+            cli_context: The current CLI context.
+            service_names: Names of the services to get the status of. If not provided, gets the status of all services.
+            extra_args: Additional arguments to be passed to the command.
 
         Returns:
             CompletedProcess: Result of the orchestrator command.
         """
+        # NOTE: extra_args not yet implemented for this orchestrator.
         # Generate the command string.
         command = [
             "helm",
@@ -814,7 +843,7 @@ class HelmOrchestrator(Orchestrator):
         """Run the given command and return the CompletedProcess.
 
         Args:
-            command (list[str]): The command to execute.
+            command: The command to execute.
 
         Returns:
             CompletedProcess: The execution result.
@@ -840,7 +869,7 @@ class HelmOrchestrator(Orchestrator):
             ["--values", "values.yaml", "--set-file", "path.to.foo=path/to/foo.yaml"...
 
         Args:
-            generated_configuration_dir (Path): Generated configuration directory form the cli object, e.g:
+            generated_configuration_dir: Generated configuration directory form the cli object, e.g:
                 `cli_context.get_generated_configuration_dir()`
 
         Returns:
@@ -886,19 +915,22 @@ class NullOrchestrator(Orchestrator):
     """
 
     def start(
-        self, cli_context: CliContext, service_name: tuple[str, ...] = None
+        self, cli_context: CliContext, service_names: tuple[str, ...] = None
     ) -> CompletedProcess:
         logger.info("NullOrchestrator has no services to start.")
         return None
 
     def shutdown(
-        self, cli_context: CliContext, service_name: tuple[str, ...] = None
+        self, cli_context: CliContext, service_names: tuple[str, ...] = None
     ) -> CompletedProcess:
         logger.info("NullOrchestrator has no services to shutdown.")
         return None
 
     def status(
-        self, cli_context: CliContext, service_name: tuple[str, ...] = None
+        self,
+        cli_context: CliContext,
+        service_names: tuple[str, ...] = None,
+        extra_args: tuple[str, ...] = None,
     ) -> CompletedProcess:
         logger.info("NullOrchestrator has no services to get the status of.")
         return None
@@ -964,8 +996,8 @@ def service_name_verifier(
     """Verify all services exist.
 
     Args:
-        service_names (tuple[str, ...]): The list of service names to check.
-        valid_service_names [List[str]]: The list of valid service names.
+        service_names: The list of service names to check.
+        valid_service_names: The list of valid service names.
 
     """
     invalid_service_names = set(service_names) - set(valid_service_names)
@@ -984,10 +1016,10 @@ def decrypt_docker_compose_files(
     """Decrypt docker-compose and docker-compose override files.
 
     Args:
-        cli_context (CliContext): The current CLI context.
-        docker_compose_file_relative_path (Path): The relative path to the docker-compose file. Path is relative to the
+        cli_context: The current CLI context.
+        docker_compose_file_relative_path: The relative path to the docker-compose file. Path is relative to the
             generated configuration directory.
-        docker_compose_override_directory_relative_path (Path): The relative path to a directory containing
+        docker_compose_override_directory_relative_path: The relative path to a directory containing
             docker-compose override files. Path is relative to the generated configuration directory.
 
     Returns:
@@ -1068,14 +1100,14 @@ def execute_compose(
     """Builds and executes a docker-compose command.
 
     Args:
-        cli_context (CliContext): The current CLI context.
-        command (Iterable[str]): The command to execute with docker-compose.
-        docker_compose_file_relative_path (Path): The relative path to the docker-compose file. Path is relative to the
+        cli_context: The current CLI context.
+        command: The command to execute with docker-compose.
+        docker_compose_file_relative_path: The relative path to the docker-compose file. Path is relative to the
             generated configuration directory.
-        docker_compose_override_directory_relative_path (Path): The relative path to a directory containing
+        docker_compose_override_directory_relative_path: The relative path to a directory containing
             docker-compose override files. Path is relative to the generated configuration directory.
-        stdin_input (str): Optional - defaults to None. String passed through to the subprocess via stdin.
-        capture_output (bool): Optional - defaults to False. True to capture stdout/stderr for the run command.
+        stdin_input: Optional - defaults to None. String passed through to the subprocess via stdin.
+        capture_output: Optional - defaults to False. True to capture stdout/stderr for the run command.
 
     Returns:
         CompletedProcess: The completed process and its exit code.